Unify PIN/UV shared secret handling
Move `PinUvAuthSharedSecret` into `pin_uv_auth_protocol` as the central protocol-agnostic wrapper and migrate its tests there. Refactor `client_pin` to use the unified API for token retrieval, PIN set/change, encryption, and auth param generation, removing duplicated v1/v2 branching and helper functions. Update call sites/imports (including getAssertion and hmac extension) and make v1 token decryption accept an immutable ciphertext slice.
