+++ src/fidokey/pin/client_pin.rs
@@ -3,14 +3,9 @@ use super::client_pin_command::Permission;
 use super::client_pin_command::SubCommand as PinCmd;
 use super::client_pin_response;
 use super::FidoKeyHid;
-use crate::crypto::rand;
-use crate::crypto::rand::SecureRandom;
 use crate::ctaphid;
 use crate::encrypt::cose;
-use crate::encrypt::enc_aes256_cbc;
-use crate::encrypt::shared_secret::SharedSecret;
-use crate::encrypt::shared_secret2::SharedSecret2;
-use crate::pin_uv_auth_protocol::{self, PinUvAuthProtocol};
+use crate::pin_uv_auth_protocol::{compute_pin_uv_auth_param, PinUvAuthSharedSecret};
 use crate::pintoken::PinToken;
 use anyhow::{anyhow, Result};
 
@@ -27,11 +22,7 @@ impl FidoKeyHid {
     pub fn create_pin_auth(&self, pin: &str, client_data_hash: &[u8]) -> Result<Vec<u8>> {
         let pin_token = self.get_pin_token(pin)?;
 
-        pin_uv_auth_protocol::compute_pin_uv_auth_param(
-            &pin_token.key,
-            client_data_hash,
-            self.pin_protocol_version,
-        )
+        compute_pin_uv_auth_param(&pin_token.key, client_data_hash, self.pin_protocol_version)
     }
 
     pub fn get_pin_token(&self, pin: &str) -> Result<PinToken> {
@@ -40,49 +31,26 @@ impl FidoKeyHid {
         }
 
         let authenticator_key_agreement = self.get_authenticator_key_agreement()?;
+        let shared_secret =
+            PinUvAuthSharedSecret::new(self.pin_protocol_version, &authenticator_key_agreement)?;
 
-        let pin_token_dec = match PinUvAuthProtocol::from_version(self.pin_protocol_version)? {
-            PinUvAuthProtocol::One => {
-                let shared_secret = SharedSecret::new(&authenticator_key_agreement)?;
-                let pin_hash_enc = shared_secret.encrypt_pin(pin)?;
+        // Get pinHashEnc
+        let pin_hash_enc = shared_secret.encrypt_pin(pin)?;
 
-                let send_payload = client_pin_command::create_payload_get_pin_token(
-                    &shared_secret.public_key,
-                    &pin_hash_enc,
-                    self.pin_protocol_version,
-                )?;
+        let send_payload = client_pin_command::create_payload_get_pin_token(
+            shared_secret.public_key(),
+            &pin_hash_enc,
+            self.pin_protocol_version,
+        )?;
 
-                let response_cbor = ctaphid::ctaphid_cbor(self, &send_payload)?;
+        let response_cbor = ctaphid::ctaphid_cbor(self, &send_payload)?;
 
-                // get pin_token (enc)
-                let mut pin_token_enc =
-                    client_pin_response::parse_cbor_client_pin_get_pin_token(&response_cbor)?;
+        // get pin_token (enc)
+        let pin_token_enc =
+            client_pin_response::parse_cbor_client_pin_get_pin_token(&response_cbor)?;
 
-                // pintoken -> dec(pintoken)
-                shared_secret.decrypt_token(&mut pin_token_enc)?
-            }
-            PinUvAuthProtocol::Two => {
-                let shared_secret = SharedSecret2::new(&authenticator_key_agreement)?;
-                let pin_hash_enc = shared_secret.encrypt_pin(pin)?;
-
-                let send_payload = client_pin_command::create_payload_get_pin_token(
-                    &shared_secret.public_key,
-                    &pin_hash_enc,
-                    self.pin_protocol_version,
-                )?;
-
-                let response_cbor = ctaphid::ctaphid_cbor(self, &send_payload)?;
-
-                // get pin_token (enc)
-                let pin_token_enc =
-                    client_pin_response::parse_cbor_client_pin_get_pin_token(&response_cbor)?;
-
-                // pintoken -> dec(pintoken)
-                shared_secret.decrypt_token(&pin_token_enc)?
-            }
-        };
-
-        Ok(pin_token_dec)
+        // pintoken -> dec(pintoken)
+        shared_secret.decrypt_token(&pin_token_enc)
     }
 
     pub fn get_pinuv_auth_token_with_permission(
@@ -95,57 +63,28 @@ impl FidoKeyHid {
         }
 
         let authenticator_key_agreement = self.get_authenticator_key_agreement()?;
+        let shared_secret =
+            PinUvAuthSharedSecret::new(self.pin_protocol_version, &authenticator_key_agreement)?;
 
-        let pin_token_dec = match PinUvAuthProtocol::from_version(self.pin_protocol_version)? {
-            PinUvAuthProtocol::One => {
-                // Get pinHashEnc
-                // - shared_secret.public_key -> platform KeyAgreement
-                let shared_secret = SharedSecret::new(&authenticator_key_agreement)?;
-                let pin_hash_enc = shared_secret.encrypt_pin(pin)?;
+        // Get pinHashEnc
+        let pin_hash_enc = shared_secret.encrypt_pin(pin)?;
 
-                // Get pin token
-                let send_payload =
-                    client_pin_command::create_payload_get_pin_uv_auth_token_using_pin_with_permissions(
-                        &shared_secret.public_key,
-                        &pin_hash_enc,
-                        permission,
-                        self.pin_protocol_version,
-                    )?;
-                let response_cbor = ctaphid::ctaphid_cbor(self, &send_payload)?;
+        // Get pin token
+        let send_payload =
+            client_pin_command::create_payload_get_pin_uv_auth_token_using_pin_with_permissions(
+                shared_secret.public_key(),
+                &pin_hash_enc,
+                permission,
+                self.pin_protocol_version,
+            )?;
+        let response_cbor = ctaphid::ctaphid_cbor(self, &send_payload)?;
 
-                // get pin_token (enc)
-                let mut pin_token_enc =
-                    client_pin_response::parse_cbor_client_pin_get_pin_token(&response_cbor)?;
+        // get pin_token (enc)
+        let pin_token_enc =
+            client_pin_response::parse_cbor_client_pin_get_pin_token(&response_cbor)?;
 
-                // pintoken -> dec(pintoken)
-                shared_secret.decrypt_token(&mut pin_token_enc)?
-            }
-            PinUvAuthProtocol::Two => {
-                // Get pinHashEnc
-                // - shared_secret.public_key -> platform KeyAgreement
-                let shared_secret = SharedSecret2::new(&authenticator_key_agreement)?;
-                let pin_hash_enc = shared_secret.encrypt_pin(pin)?;
-
-                // Get pin token
-                let send_payload =
-                    client_pin_command::create_payload_get_pin_uv_auth_token_using_pin_with_permissions(
-                        &shared_secret.public_key,
-                        &pin_hash_enc,
-                        permission,
-                        self.pin_protocol_version,
-                    )?;
-                let response_cbor = ctaphid::ctaphid_cbor(self, &send_payload)?;
-
-                // get pin_token (enc)
-                let pin_token_enc =
-                    client_pin_response::parse_cbor_client_pin_get_pin_token(&response_cbor)?;
-
-                // pintoken -> dec(pintoken)
-                shared_secret.decrypt_token(&pin_token_enc)?
-            }
-        };
-
-        Ok(pin_token_dec)
+        // pintoken -> dec(pintoken)
+        shared_secret.decrypt_token(&pin_token_enc)
     }
 
     pub fn set_new_pin_cmd(&self, pin: &str) -> Result<()> {
@@ -161,32 +100,17 @@ impl FidoKeyHid {
         let key_agreement =
             client_pin_response::parse_cbor_client_pin_get_keyagreement(&response_cbor)?;
 
-        // get public_key, pin_auth, new_pin_enc
-        let (public_key, pin_auth, new_pin_enc) =
-            match PinUvAuthProtocol::from_version(self.pin_protocol_version)? {
-                PinUvAuthProtocol::One => {
-                    let shared_secret = SharedSecret::new(&key_agreement)?;
+        let shared_secret = PinUvAuthSharedSecret::new(self.pin_protocol_version, &key_agreement)?;
 
-                    let new_pin_enc = create_new_pin_enc(&shared_secret, pin)?;
+        // newPinEnc: encrypt(shared secret, paddedPin)
+        let new_pin_enc = shared_secret.encrypt(&padding_pin_64(pin)?)?;
 
-                    let pin_auth = create_pin_auth_for_set_pin(&shared_secret, &new_pin_enc)?;
-
-                    (shared_secret.public_key, pin_auth, new_pin_enc)
-                }
-                PinUvAuthProtocol::Two => {
-                    let shared_secret = SharedSecret2::new(&key_agreement)?;
-
-                    let new_pin_enc = create_new_pin_enc2(&shared_secret, pin)?;
-
-                    let pin_auth = create_pin_auth_for_set_pin2(&shared_secret, &new_pin_enc)?;
-
-                    (shared_secret.public_key, pin_auth, new_pin_enc)
-                }
-            };
+        // pinUvAuthParam: authenticate(shared secret, newPinEnc)
+        let pin_auth = shared_secret.authenticate(&new_pin_enc)?;
 
         // set new pin
         let send_payload = client_pin_command::create_payload_set_pin(
-            &public_key,
+            shared_secret.public_key(),
             &pin_auth,
             &new_pin_enc,
             self.pin_protocol_version,
@@ -213,59 +137,24 @@ impl FidoKeyHid {
         let key_agreement =
             client_pin_response::parse_cbor_client_pin_get_keyagreement(&response_cbor)?;
 
-        // get public_key, pin_auth, new_pin_enc, current_pin_hash_enc
-        let (public_key, pin_auth, new_pin_enc, current_pin_hash_enc) =
-            match PinUvAuthProtocol::from_version(self.pin_protocol_version)? {
-                PinUvAuthProtocol::One => {
-                    let shared_secret = SharedSecret::new(&key_agreement)?;
+        let shared_secret = PinUvAuthSharedSecret::new(self.pin_protocol_version, &key_agreement)?;
 
-                    let current_pin_hash_enc = shared_secret.encrypt_pin(current_pin)?;
+        // https://fidoalliance.org/specs/fido-v2.1-ps-20210615/fido-client-to-authenticator-protocol-v2.1-ps-20210615.html#changingExistingPin
+        // 6.5.5.6. Changing existing PIN
 
-                    let new_pin_enc = create_new_pin_enc(&shared_secret, new_pin)?;
+        // pinHashEnc: encrypt(shared secret, LEFT(SHA-256(curPin), 16))
+        let current_pin_hash_enc = shared_secret.encrypt_pin(current_pin)?;
 
-                    let pin_auth = create_pin_auth_for_change_pin(
-                        &shared_secret,
-                        &new_pin_enc,
-                        &current_pin_hash_enc,
-                    )?;
+        // newPinEnc: encrypt(shared secret, paddedPin)
+        let new_pin_enc = shared_secret.encrypt(&padding_pin_64(new_pin)?)?;
 
-                    (
-                        shared_secret.public_key,
-                        pin_auth,
-                        new_pin_enc,
-                        current_pin_hash_enc.into(),
-                    )
-                }
-                PinUvAuthProtocol::Two => {
-                    // https://fidoalliance.org/specs/fido-v2.1-ps-20210615/fido-client-to-authenticator-protocol-v2.1-ps-20210615.html#changingExistingPin
-                    // 6.5.5.6. Changing existing PIN
-
-                    let shared_secret = SharedSecret2::new(&key_agreement)?;
-
-                    // 4. pinHashEnc: The result of calling encrypt(shared secret, LEFT(SHA-256(curPin), 16)).
-                    let current_pin_hash_enc = shared_secret.encrypt_pin(current_pin)?;
-
-                    // 5. newPinEnc: the result of calling encrypt(shared secret, paddedPin) where paddedPin is newPin padded on the right with 0x00 bytes to make it 64 bytes long. (Since the maximum length of newPin is 63 bytes, there is always at least one byte of padding.)
-                    let new_pin_enc = create_new_pin_enc2(&shared_secret, new_pin)?;
-
-                    // 6. pinUvAuthParam: the result of calling authenticate(shared secret, newPinEnc || pinHashEnc).
-                    let pin_auth = create_pin_auth_for_change_pin2(
-                        &shared_secret,
-                        &new_pin_enc,
-                        &current_pin_hash_enc,
-                    )?;
-
-                    (
-                        shared_secret.public_key,
-                        pin_auth,
-                        new_pin_enc,
-                        current_pin_hash_enc,
-                    )
-                }
-            };
+        // pinUvAuthParam: authenticate(shared secret, newPinEnc || pinHashEnc)
+        let mut message = new_pin_enc.clone();
+        message.extend_from_slice(&current_pin_hash_enc);
+        let pin_auth = shared_secret.authenticate(&message)?;
 
         let send_payload = client_pin_command::create_payload_change_pin(
-            &public_key,
+            shared_secret.public_key(),
             &pin_auth,
             &new_pin_enc,
             &current_pin_hash_enc,
@@ -276,46 +165,6 @@ impl FidoKeyHid {
 
         Ok(())
     }
-}
-
-fn create_pin_auth_for_set_pin(
-    shared_secret: &SharedSecret,
-    new_pin_enc: &[u8],
-) -> Result<Vec<u8>> {
-    pin_uv_auth_protocol::compute_pin_uv_auth_param(&shared_secret.secret, new_pin_enc, 1)
-}
-
-fn create_pin_auth_for_set_pin2(
-    shared_secret: &SharedSecret2,
-    new_pin_enc: &[u8],
-) -> Result<Vec<u8>> {
-    pin_uv_auth_protocol::compute_pin_uv_auth_param(&shared_secret.secret[0..32], new_pin_enc, 2)
-}
-
-fn create_pin_auth_for_change_pin(
-    shared_secret: &SharedSecret,
-    new_pin_enc: &[u8],
-    current_pin_hash_enc: &[u8],
-) -> Result<Vec<u8>> {
-    // source data
-    let mut message = vec![];
-    message.append(&mut new_pin_enc.to_vec());
-    message.append(&mut current_pin_hash_enc.to_vec());
-
-    pin_uv_auth_protocol::compute_pin_uv_auth_param(&shared_secret.secret, &message, 1)
-}
-
-fn create_pin_auth_for_change_pin2(
-    shared_secret: &SharedSecret2,
-    new_pin_enc: &[u8],
-    current_pin_hash_enc: &[u8],
-) -> Result<Vec<u8>> {
-    // source data
-    let mut message = vec![];
-    message.append(&mut new_pin_enc.to_vec());
-    message.append(&mut current_pin_hash_enc.to_vec());
-
-    pin_uv_auth_protocol::compute_pin_uv_auth_param(&shared_secret.secret[0..32], &message, 2)
 }
 
 fn padding_pin_64(pin: &str) -> Result<Vec<u8>> {
@@ -334,34 +183,4 @@ fn padding_pin_64(pin: &str) -> Result<Vec<u8>> {
     }
 
     Ok(bpin64)
-}
-
-// newPinEnc: AES256-CBC(sharedSecret, IV = 0, newPin)
-fn create_new_pin_enc(shared_secret: &SharedSecret, new_pin: &str) -> Result<Vec<u8>> {
-    let new_pin_64 = padding_pin_64(new_pin)?;
-
-    let new_pin_enc = enc_aes256_cbc::encrypt_message(&shared_secret.secret, &new_pin_64);
-
-    Ok(new_pin_enc)
-}
-fn create_new_pin_enc2(shared_secret: &SharedSecret2, new_pin: &str) -> Result<Vec<u8>> {
-    let new_pin_64 = padding_pin_64(new_pin)?;
-
-    let aes_key: &[u8; 32] = shared_secret.secret[32..].try_into()?;
-
-    // Let iv be a 16-byte, random bytestring.
-    let mut iv = [0u8; 16];
-    let rng = rand::SystemRandom::new();
-    rng.fill(&mut iv)
-        .map_err(|_| anyhow!("Failed to generate random IV"))?;
-
-    let ciphertext = enc_aes256_cbc::encrypt_message_with_iv(aes_key, &iv, &new_pin_64);
-
-    // Return iv || ct.
-    // Concatenate iv and ct(ciphertext)
-    let mut new_pin_enc = vec![];
-    new_pin_enc.extend_from_slice(&iv);
-    new_pin_enc.extend_from_slice(&ciphertext);
-
-    Ok(new_pin_enc)
 }
