+++ src/pin_uv_auth_protocol.rs
@@ -1,4 +1,8 @@
+use crate::encrypt::cose::CoseKey;
 use crate::encrypt::enc_hmac_sha_256;
+use crate::encrypt::shared_secret::SharedSecret;
+use crate::encrypt::shared_secret2::SharedSecret2;
+use crate::pintoken::PinToken;
 use anyhow::{anyhow, Result};
 
 /// The CTAP2 PIN/UV Auth Protocol negotiated with the authenticator.
@@ -41,5 +45,151 @@ pub fn compute_pin_uv_auth_param(
     match protocol {
         PinUvAuthProtocol::One => Ok(sig[0..16].to_vec()),
         PinUvAuthProtocol::Two => Ok(sig),
+    }
+}
+
+/// A PIN/UV Auth Protocol shared secret, in whichever protocol version was negotiated.
+///
+/// Centralizes the Protocol One vs Protocol Two split (key material, encrypt/decrypt,
+/// and HMAC truncation rules) so callers that need a shared secret — for the
+/// pinUvAuthToken flow, PIN set/change, or an extension like hmac-secret — don't have
+/// to re-implement the split themselves.
+#[derive(Debug, Clone)]
+pub enum PinUvAuthSharedSecret {
+    One(SharedSecret),
+    Two(SharedSecret2),
+}
+
+impl Default for PinUvAuthSharedSecret {
+    fn default() -> Self {
+        Self::One(SharedSecret::default())
+    }
+}
+
+impl PinUvAuthSharedSecret {
+    pub fn new(pin_protocol_version: u8, key_agreement: &CoseKey) -> Result<Self> {
+        match PinUvAuthProtocol::from_version(pin_protocol_version)? {
+            PinUvAuthProtocol::One => Ok(Self::One(SharedSecret::new(key_agreement)?)),
+            PinUvAuthProtocol::Two => Ok(Self::Two(SharedSecret2::new(key_agreement)?)),
+        }
+    }
+
+    pub fn public_key(&self) -> &CoseKey {
+        match self {
+            Self::One(s) => &s.public_key,
+            Self::Two(s) => &s.public_key,
+        }
+    }
+
+    pub fn pin_protocol_version(&self) -> u8 {
+        match self {
+            Self::One(_) => PinUvAuthProtocol::One.version(),
+            Self::Two(_) => PinUvAuthProtocol::Two.version(),
+        }
+    }
+
+    /// The key used for authenticate()/HMAC-SHA-256. Protocol Two discards the
+    /// AES-key half of the shared secret, keeping only the HMAC-key half.
+    fn auth_key(&self) -> &[u8] {
+        match self {
+            Self::One(s) => &s.secret,
+            Self::Two(s) => &s.secret[0..32],
+        }
+    }
+
+    /// encrypt(key, demPlaintext) → ciphertext, per the negotiated protocol version.
+    pub fn encrypt(&self, dem_plaintext: &[u8]) -> Result<Vec<u8>> {
+        match self {
+            Self::One(s) => Ok(s.encrypt_raw(dem_plaintext)),
+            Self::Two(s) => s.encrypt(dem_plaintext),
+        }
+    }
+
+    /// decrypt(key, demCiphertext) → plaintext, per the negotiated protocol version.
+    pub fn decrypt(&self, dem_cipher_text: &[u8]) -> Result<Vec<u8>> {
+        match self {
+            Self::One(s) => Ok(s.decrypt_raw(dem_cipher_text)),
+            Self::Two(s) => s.decrypt(dem_cipher_text),
+        }
+    }
+
+    /// authenticate(key, message) → pinUvAuthParam, per the negotiated protocol version.
+    pub fn authenticate(&self, message: &[u8]) -> Result<Vec<u8>> {
+        compute_pin_uv_auth_param(self.auth_key(), message, self.pin_protocol_version())
+    }
+
+    /// pinHashEnc/pinUvAuthToken encryption of a raw PIN string, per the negotiated
+    /// protocol version's PIN-hashing rule (LEFT(SHA-256(pin), 16), then encrypt()).
+    pub fn encrypt_pin(&self, pin: &str) -> Result<Vec<u8>> {
+        match self {
+            Self::One(s) => Ok(s.encrypt_pin(pin)?.to_vec()),
+            Self::Two(s) => s.encrypt_pin(pin),
+        }
+    }
+
+    /// Decrypt an authenticator-provided pinUvAuthToken ciphertext into a PinToken.
+    pub fn decrypt_token(&self, dem_cipher_text: &[u8]) -> Result<PinToken> {
+        match self {
+            Self::One(s) => s.decrypt_token(dem_cipher_text),
+            Self::Two(s) => s.decrypt_token(dem_cipher_text),
+        }
+    }
+}
+
+#[cfg(test)]
+mod pin_uv_auth_shared_secret_tests {
+    use super::*;
+
+    fn one() -> PinUvAuthSharedSecret {
+        PinUvAuthSharedSecret::One(SharedSecret {
+            secret: [3u8; 32],
+            public_key: CoseKey::default(),
+        })
+    }
+
+    fn two() -> PinUvAuthSharedSecret {
+        let mut secret = [0u8; 64];
+        secret[..32].copy_from_slice(&[1u8; 32]); // HMAC-key half
+        secret[32..].copy_from_slice(&[2u8; 32]); // AES-key half
+        PinUvAuthSharedSecret::Two(SharedSecret2 {
+            secret,
+            public_key: CoseKey::default(),
+        })
+    }
+
+    #[test]
+    fn test_pin_uv_auth_shared_secret_protocol_versions() {
+        assert_eq!(one().pin_protocol_version(), 1);
+        assert_eq!(two().pin_protocol_version(), 2);
+    }
+
+    #[test]
+    fn test_pin_uv_auth_shared_secret_encrypt_decrypt_roundtrip() {
+        let salt = [9u8; 32];
+
+        for shared_secret in [one(), two()] {
+            let salt_enc = shared_secret.encrypt(&salt).unwrap();
+            let decrypted = shared_secret.decrypt(&salt_enc).unwrap();
+            assert_eq!(decrypted, salt);
+        }
+    }
+
+    #[test]
+    fn test_pin_uv_auth_shared_secret_authenticate_length() {
+        let message = [5u8; 32];
+
+        // Protocol One truncates to 16 bytes.
+        let param_one = one().authenticate(&message).unwrap();
+        assert_eq!(param_one.len(), 16);
+
+        // Protocol Two returns the full 32-byte HMAC-SHA-256 output.
+        let param_two = two().authenticate(&message).unwrap();
+        assert_eq!(param_two.len(), 32);
+    }
+
+    #[test]
+    fn test_pin_uv_auth_shared_secret_new_rejects_unknown_protocol() {
+        let result = PinUvAuthSharedSecret::new(3, &CoseKey::default());
+        assert!(result.is_err());
     }
 }
